Add createSeekAttachmentMiddleware
This configures a proxy handler for attachment downloads. The major
difference compared to the original implementation is that this has been
rebased on `node-fetch` and passes a readable stream through to Koa.

I’ve tried to reuse the same `GetPartnerToken` definition here but I
think the cracks are starting to show. Once the middlewares have settled
a little, we can revisit its signature.