Add createPartnerWebhookMiddleware stub (#16)
This only formally covers input validation and signature verification,
though storage could be implemented through the callback.

It also wires up the example frontend and backend with a new feature:
the frontend can POST a fake webhook. If we can’t figure out a nice way
to get the webhook interaction working in a self-contained manner, we
may just support some form of seeding. This could be triggered at
startup and/or by user interaction.