Add createBrowserTokenMiddleware
This is better practice than using the SeekGraph proxy as the access is
scoped to a hirer, as it would be in an actual third-party system. The
interface is a bit rough but hopefully the core functionality is there;
I'll follow this up with an actual implementation in the example
frontend.