Craft artisanal deployment
This drops the Storybook deployer in favour of a simple script with
plenty of hardcoding. This allows us to trivially customise the
directory structure of the `gh-pages` branch and add in our mock site.

<https://seek-oss.github.io/wingman>

I initially thought of running a script after the Storybook deployer,
but unfortunately it deletes its working directory after deployment.