Use type-only imports and exports
Our frontend build is currently failing. While investigating the issue I
noticed a lot of warnings in our `sku build --debug` output when Webpack
is unable to locate an import that has been stripped as it's a
TypeScript type.

This is admittedly an overreaction. It resolves those warnings but not
the actual failure.