Codegen TypeScript typeDefs (#532)
Previously, we generated a `.graphql` source file and pulled it into
JavaScript land with a Webpack loader. This vendors a simple plugin to
generate a TypeScript source file from the get-go, which is easier to
work with and opens the door for us to share `ApolloMockProvider` with
consumers without additional Webpack configuration.