Control GitHub links with `showStorybookAction`
When this is `false`/`undefined` we still show the "View on GitHub" link
which is unlikely what the consumer wants. This repurposes the flag to
control both the storybook & GitHub links as a group.

This also lets us tighten the types on our internal
`MockComponentActions` to ensure the mock component is providing all the
relevant links.